pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my_grc"
version = "0.3.0"
description = "Workbook-first GRC platform for unified controls, scoped implementation, evidence, assessments, and governed AI assistance."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "boto3>=1.34.0",
  "PyYAML>=6.0.1",
  "SQLAlchemy>=2.0.29",
  "requests>=2.31.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "python-dateutil>=2.9.0",
  "cryptography>=43.0.0",
  "keyring>=25.3.0",
  "pandas>=2.2.2",
  "openpyxl>=3.1.5",
  "streamlit>=1.38.0",
]

[project.optional-dependencies]
browser-evidence = [
  "selenium>=4.29.0",
]
dev = [
  "pytest>=8.3.4",
  "pip-audit>=2.8.0",
]

[project.scripts]
aws-local-audit = "aws_local_audit.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aws_local_audit = ["templates/frameworks/*.yaml"]
